# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "POS API Documentation",
        "description": "API Documentation for Point of Sale System",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/cart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Get user's cart",
                "operationId": "0197bc7dd36ee12bfc09b13a30c959bb",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "description": "Number of products to skip",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of products to return",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 30
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cart data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "carts": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "products": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "title": {
                                                                "type": "string"
                                                            },
                                                            "price": {
                                                                "type": "number",
                                                                "format": "float"
                                                            },
                                                            "quantity": {
                                                                "type": "integer"
                                                            },
                                                            "total": {
                                                                "type": "number",
                                                                "format": "float"
                                                            },
                                                            "discountPercentage": {
                                                                "type": "number",
                                                                "format": "float"
                                                            },
                                                            "discountedTotal": {
                                                                "type": "number",
                                                                "format": "float"
                                                            },
                                                            "thumbnail": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "total": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "discountedTotal": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "userId": {
                                                    "type": "integer"
                                                },
                                                "totalProducts": {
                                                    "type": "integer"
                                                },
                                                "totalQuantity": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "total": {
                                            "type": "integer"
                                        },
                                        "skip": {
                                            "type": "integer"
                                        },
                                        "limit": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cart/add": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Add items to cart",
                "operationId": "1fd304ca79daacf6570caa5e412bccc7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "userId",
                                    "products"
                                ],
                                "properties": {
                                    "userId": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Products added to cart successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "cart": {
                                            "type": "object"
                                        },
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "type": "number",
                                            "format": "float"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}